requirements.txt
stream_django
djangorestframework
pycrypto
